feat: return specific error when authentication with password fails
Authentication failure may be handled by the caller e.g. by asking the
user to update or provide the password like in usql, so it is valuable to
expose a way to identify such failures in the API.

The implementation avoids exposing thrift to the sasl package.
sasl implementation layers will likely be refactored and simplified
after #29.

Testing Done: make test; new integration test